Walk ARM Procedure Call Standard frames for profiling, where available.
This needs more work, but it's a proof-of-concept and it basically
works.

Also this is not quite APCS, as it ignores the saved PC field (as
distinct from the saved LR), but: Linux perf(1) does this too, and I'm
exploiting this property in my GCC change to make this work efficiently
for Thumb2.