Bug 888578: extend ARM masm's floatArgsInGPR workaround to memory loads
The move resolver doesn't understand using a pair of registers, which
we have to do on ARM softfp to pass doubles in the integer registers,
if they occur early enough in the argument list.  We had a workaround
(see uses of floatArgsInGPR) for when the argument is moved from a float
register; this change extends it to handle loading from memory.